services/backend/app/requirements.txt

fastapi==0.100.0
pydantic>=1.8.0,<2.0.0
uvicorn==0.22.0

python-docx
reportlab
gdown